fix gc tests running in vanilla pipeline
Conflicts:
	go.mod
	go.sum
