+++ tests/e2e/fullsync_test_for_block_volume.go
@@ -690,114 +690,4 @@ var _ bool = ginkgo.Describe("[csi-block-vanilla] full-sync-test", func() {
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	})
 
-	/*
-		Create a PVC using any replicated storage class from the SV.
-		Wait for PVC to be in Bound phase
-		Break connection to SV
-		Update PV and PVC labels
-		Re-establish connection to SV
-		Sleep double the Full Sync interval
-		Verify CnsVolumeMetadata CRD is updated
-		Verify entry is updated in CNS
-		Delete PVC
-	*/
-	ginkgo.It("[csi-guest] Verify CNS volume is synced with updated GC PV & PVC labels after SVC connection is restored", func() {
-		var err error
-		var sc *storagev1.StorageClass
-		ctx, cancel := context.WithCancel(context.Background())
-		defer cancel()
-
-		ginkgo.By("CNS_TEST: Running for GC setup")
-		ginkgo.By("Creating Storage Class and PVC")
-		scParameters[svStorageClassName] = storagePolicyName
-		sc, pvc, err := createPVCAndStorageClass(client, namespace, nil, scParameters, "", nil, "", false, "")
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-		defer func() {
-			err := client.StorageV1().StorageClasses().Delete(ctx, sc.Name, *metav1.NewDeleteOptions(0))
-			gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		}()
-
-		ginkgo.By(fmt.Sprintf("Waiting for claim %s to be in bound phase", pvc.Name))
-		pvs, err := fpv.WaitForPVClaimBoundPhase(client, []*v1.PersistentVolumeClaim{pvc}, framework.ClaimProvisionTimeout)
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		gomega.Expect(pvs).NotTo(gomega.BeEmpty())
-		pv := pvs[0]
-		volumeID := pv.Spec.CSI.VolumeHandle
-		svcVolumeID := getVolumeIDFromSupervisorCluster(volumeID)
-		gomega.Expect(svcVolumeID).NotTo(gomega.BeEmpty())
-		ginkgo.By(fmt.Sprintf("SVC volume ID for claim %s is %s", volumeID, svcVolumeID))
-		defer func() {
-			err := fpv.DeletePersistentVolumeClaim(client, pvc.Name, namespace)
-			gomega.Expect(err).NotTo(gomega.HaveOccurred())
-			err = e2eVSphere.waitForCNSVolumeToBeDeleted(pv.Spec.CSI.VolumeHandle)
-			gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		}()
-
-		fysncInterval := string(int((fullSyncWaitTime - 20) / 120))
-		ginkgo.By(fmt.Sprintf("Reducing full sync interval to %v mins for the test...", fysncInterval))
-		_ = updateCSIDeploymentTemplateFullSyncInterval(client, fysncInterval, csiSystemNamespace)
-		defer func() {
-			ginkgo.By(fmt.Sprintf("Resetting full sync interval to %v mins", defaultFullSyncIntervalInMin))
-			_ = updateCSIDeploymentTemplateFullSyncInterval(client, defaultFullSyncIntervalInMin, csiSystemNamespace)
-		}()
-
-		// break connection with SVC
-		ginkgo.By("Scaling down the csi driver to zero replica")
-		deployment := updateDeploymentReplica(client, 0, vSphereCSIControllerPodNamePrefix, csiSystemNamespace)
-		ginkgo.By(fmt.Sprintf("Successfully scaled down the csi driver deployment:%s to zero replicas", deployment.Name))
-
-		pvLabels := make(map[string]string)
-		pvLabels["pvLabelKeyFsync"] = "pv-label-Value-fsync"
-
-		pvcLabels := make(map[string]string)
-		pvcLabels["pvcLabelKeyFsync"] = "pvc-label-Value-fsync"
-
-		ginkgo.By(fmt.Sprintf("Updating labels %+v for pvc %s in namespace %s", pvcLabels, pvc.Name, pvc.Namespace))
-		pvc, err = client.CoreV1().PersistentVolumeClaims(namespace).Get(ctx, pvc.GetName(), metav1.GetOptions{})
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		pvc.Labels = pvcLabels
-		pvc, err = client.CoreV1().PersistentVolumeClaims(namespace).Update(ctx, pvc, metav1.UpdateOptions{})
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-		ginkgo.By(fmt.Sprintf("Updating labels %+v for pv %s in namespace %s", pvLabels, pv.Name, namespace))
-		pv, err = client.CoreV1().PersistentVolumes().Get(ctx, pv.GetName(), metav1.GetOptions{})
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		pv.Labels = pvLabels
-		pv, err = client.CoreV1().PersistentVolumes().Update(ctx, pv, metav1.UpdateOptions{})
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-		// re-establish connection with SVC
-		ginkgo.By("Scaling up the csi driver to one replica")
-		deployment = updateDeploymentReplica(client, 1, vSphereCSIControllerPodNamePrefix, csiSystemNamespace)
-		ginkgo.By(fmt.Sprintf("Successfully scaled up the csi driver deployment:%s to one replica", deployment.Name))
-
-		ginkgo.By(fmt.Sprintf("Sleeping for %v seconds to allow full sync to finish", fullSyncWaitTime))
-		time.Sleep(time.Duration(fullSyncWaitTime) * time.Second)
-
-		fmt.Println("PVC name in SV", volumeID)
-		pvcUID := string(pvc.GetUID())
-		fmt.Println("PVC UUID in GC", pvcUID)
-		gcClusterID := strings.Replace(volumeID, pvcUID, "", -1)
-
-		fmt.Println("gcClusterId", gcClusterID)
-		pvUID := string(pv.UID)
-		fmt.Println("PV uuid", pvUID)
-
-		// check pvc label update in CRD
-		verifyEntityReferenceInCRDInSupervisor(ctx, f, volumeID, crdCNSVolumeMetadatas, crdVersion, crdGroup, true, volumeID, true, pvcLabels, true)
-
-		// check pv label update in CRD
-		verifyEntityReferenceInCRDInSupervisor(ctx, f, gcClusterID+pvUID, crdCNSVolumeMetadatas, crdVersion, crdGroup, true, volumeID, true, pvLabels, true)
-
-		ginkgo.By(fmt.Sprintf("Waiting for labels %+v to be updated for pvc %s in namespace %s", pvcLabels, pvc.Name, pvc.Namespace))
-		err = e2eVSphere.waitForLabelsToBeUpdated(svcVolumeID, pvcLabels, string(cnstypes.CnsKubernetesEntityTypePVC), pvc.Name, pvc.Namespace)
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-		ginkgo.By(fmt.Sprintf("Waiting for labels %+v to be updated for pv %s", pvLabels, pv.Name))
-		err = e2eVSphere.waitForLabelsToBeUpdated(svcVolumeID, pvLabels, string(cnstypes.CnsKubernetesEntityTypePV), pv.Name, pv.Namespace)
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-
-	})
-
 })
